Multi-pronged push
Split the Penicillin file

Tested textures (to be pushed by Axi), fixed XML to work with a given structure. In summary, add a Mature & an Immature folder to each plant, put them in that directory, bam. Works like a charm, structured decently. The graphics can only be pointed to a specific _folder_ so I can't really specify. Can use *A.png, *B.png, but this is more obvious, if more folder depth.