CSProj Edits
Added my hintPath back. I think there are branches without, and when it comes to merging to dev it's overlooked. Just a nuisance to me, but not much to keep on the dev branch.

Removed the Post Build events that are (guaranteed) spewing files onto your hard drive every time you build.