Generalizing the Extractable Comp to be generic
generalized the Extractable Comp in case we ever get squirrely and want to use it elsewhere (platelets, RB Cells, etc)

Now named RawFluidExtractable. JobDriver, WorkGiver, WorkGiverDef, and JobDef all still need to be Fluid specific (e.g. - insulin) because it's what will appear, and each fluid will need their own.

Included a new field/variable for the DefName to simplify the data. For some reason, RW doesn't like multiple '_'s in data in the debug viewer under the SaveKey() called function. Given the amount of time these pigs will take to accumulate the proper amount of insulin, I didn't want to give it a chance to lose progress. May be able to remove if we can prove it doesn't lose it and that it's just the variable viewer getting weird. Testing shows that 3 '_'s is just a no go in the viewer. Still not sure about the cause, but a workaround is the rawFluidDefName, and it works. A little extra, redundant info but it works. It's better than assuming how the defName is formatted or should be formatted and computing that. May, in future, change the rawFluidDefName to saveKeyString to make it more clear the intent behind it.

Made the SaveKey() and the CompInspectStringExtra() dynamic to pull the rawFluidDefName passed in via the patch XML so there shouldn't be multi-fluid conflicts, should that be a thing.